+++ CHANGELOG.md
@@ -296,6 +296,7 @@ Internal:
 * Changes to how message arguments are represented.
 * Fixed: The default 'missing argument' error message was not being used. However, this issue probably remained unnoticed until now, because most of the commands either use more specific missing argument error messages, or provide fallbacks when no arguments are provided.
 * Minor improvements to normalizing and matching enum names (used by command arguments, config settings, shopkeeper data, etc.).
+* Added a SoundEffect class to represent sound effects inside the config.
 * Minor other internal code refactoring.
 
 Migration notes: